Fix bug with map visibility on some levels.
The bug was that GameMap3D::calculateLevelRegions()
was iterating over an unordered Map collection with
after grouping the MapVoxels by depth. This resulted
in the ordering of the levelRegions list being off
sometimes. The solution was to switch to streaming
the map entries, reverse sort the stream by depth,
and then add the entries to a LinkedHashMap.